parsing code into more understandble chucks
